Fixed time, sidebar and percentage bugs
Fixed bugs on time, sidebar and percentage with same time to take to reach destined number. Fixed some CSS, like font sizes to ensure consistency for the elements.
Time on the last edit on the card was partially wrong, side bar wouldn't let user select some text now user can, with this the side bar on the pc can only be open when the mouse is bellow 124px, in mobile continues the same as before.