Updated menu and lockable sidebar
Changed:
The menu is better now.
Fixed:
Bugs when typing spaces on list name, and the system was not finding.
Addition:
In the menu for each list there is a percentage and a text for each list
More sophisticated validation for list's name and description
Now you can lock and unlock the sidebar
Now on closing the edit list menu it automatically selects the open config button.