feat(drag): automatic scrolling near borders
When a user drags an item close to the top or bottom of the page on mobile or PC, the page will automatically scroll up or down.
This enhancement allows users to continue dragging without needing to drop the item, scroll, and then drag again, thereby improving the overall user experience.